Add path argument to recombine.chunks func
Also switch to file.path()to handle file paths because it will automatically apply the correct path separator for your system. Additionally, add an if else statement to check if a directory exists, and if not, create the directory. This makes the function less reliant on if the directories exist.