skip first two lines from each subject text file
IrFanView added text to first two lines when writing tagged photos to a text file.
This caused some problems with this script because it was expecting file paths on these first two lines.
However, to implement this fix, if else statement needs to be used to detect if text file is empty.