type check date and time input and write out as strings
Excel stores date and times as doubles, which makes it difficult to work with these in R. However, some dates and or times are also being stored as strings inside of the xlsm file. I used some type checking to switch between implementation methods to convert to datetime. Then I write the datetimes out as strings so they are more easily used in R (they display exactly the same in Excel, but behind the scenes, the Excel columns are actually formatted as "Text".